+++ app/lib/theme/widgets/FileWidget.tsx
@@ -59,6 +59,8 @@ const FileWidget: React.FC<FileWidgetProps> = ({
     'primary') as string;
   const templateValidate =
     (uiSchema['ui:options']?.templateValidate as boolean) ?? false;
+  const showValidationToast =
+    (uiSchema['ui:options']?.showValidationMessage as boolean) ?? false;
   const templateNumber =
     (uiSchema['ui:options']?.templateNumber as number) ?? 0;
   const isFiles = value?.length > 0;
@@ -167,7 +169,7 @@ const FileWidget: React.FC<FileWidgetProps> = ({
         : 'Total eligible costs and Total project costs data';
     const message =
       type === 'success'
-        ? `Template ${templateNumber} validation successful, new values for ${fields} data in the application will update upon 'Save'`
+        ? `Template ${templateNumber} validation successful, new values for ${fields} in the application will update upon 'Save'`
         : `Template ${templateNumber} validation failed: ${files.join(', ')} did not validate due to formatting issues. ${fields} in the application will not update.`;
 
     showToast(message, type, 100000000);
@@ -214,7 +216,7 @@ const FileWidget: React.FC<FileWidgetProps> = ({
       span.setStatus('ok');
     }
 
-    if (templateValidate) {
+    if (templateValidate && showValidationToast) {
       if (validationErrors.length > 0) {
         showToastMessage(
           validationErrors.map(