+++ app/pages/applicantportal/form/[id]/success.tsx
@@ -29,6 +29,7 @@ const getSuccessQuery = graphql`
       intakeByIntakeId {
         ccbcIntakeNumber
         closeTimestamp
+        rollingIntake
       }
     }
 
@@ -59,8 +60,11 @@ const Success = ({
   };
 
   const projectName = applicationByRowId?.projectName;
-  const { ccbcIntakeNumber, closeTimestamp } =
-    applicationByRowId?.intakeByIntakeId || {};
+  const {
+    ccbcIntakeNumber,
+    closeTimestamp,
+    rollingIntake = false,
+  } = applicationByRowId?.intakeByIntakeId || {};
 
   const dateUpdatedAt =
     typeof applicationByRowId.updatedAt === 'string'
@@ -84,10 +88,12 @@ const Success = ({
             {` ${getDateString(dateUpdatedAt)}`} at{' '}
             {` ${getTimeString(dateUpdatedAt)}`}.
           </div>
-          <div>
-            You can edit this application until the intake closes on{' '}
-            {getDateString(dateCloseTimestamp)}
-          </div>
+          {!rollingIntake && (
+            <div>
+              You can edit this application until the intake closes on{' '}
+              {getDateString(dateCloseTimestamp)}
+            </div>
+          )}
         </StyledSection>
         <Link href="/applicantportal/dashboard" passHref>
           <Button>Return to dashboard</Button>